update masterkey class
